% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTBinteRface-package.R
\docType{package}
\name{PTBinteRface-package}
\alias{PTBinteRface-package}
\alias{PTBinteRface}
\title{Interface to the paratuberculosis model contained in the PTBinraeR package}
\description{
This is a wrapper package for INRAE's PTB model. The model can be run using the run_ptb_model function. See the vignette for more information.
}
\details{
This is a wrapper package for INRAE's PTB model.  The model can be run using the run_ptb_model function.  See the vignette for more information.
}
\author{
\strong{Maintainer}: Matt Denwood \email{md@sund.ku.dk}

Authors:
\itemize{
  \item Sandie Arnoux \email{sandie.arnoux@inrae.fr}
}

}
\keyword{internal}
