% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ptb_model.R
\name{run_ptb_model}
\alias{run_ptb_model}
\title{Wrapper function to run the Para TB model}
\usage{
run_ptb_model(
  runsNb = 1L,
  calfExposure = 1,
  cullingRateIh = 1/26,
  testSensitivityItIl = 0.15,
  testSensitivityIm = 0.47,
  testSensitivityIh = 0.71,
  testSpecificity = 1,
  startGrazing = 14L,
  endGrazing = 46L,
  ageAtWeaning = 10L,
  ageAtGrazing = 26L,
  ageYoungHeifer = 52L,
  ageHeifer = 91L,
  ageAtFirstCalving = 130L,
  noInfection = FALSE,
  InitialInfection = PTBinraeR::InitialInfection,
  BirthFemaleEvents = PTBinraeR::BirthFemaleEvents,
  CullingRates = PTBinraeR::CullingRates,
  Headcounts = PTBinraeR::Headcounts,
  Network = PTBinraeR::Network,
  seed = sample(.Machine$integer.max, 1)
)
}
\arguments{
\item{runsNb}{number of repetitions (default=1L)}

\item{calfExposure}{calf rearing improvement is defined as a reduced exposure to the general environment by varying parameter from 1.0 to 0.35 (default=1.0)}

\item{cullingRateIh}{culling rate of detected Ih animals (default=1.0/26.0)}

\item{testSensitivityItIl}{test sensitivity of low positive animals. (default=0.15)}

\item{testSensitivityIm}{test sensitivity of moderately positive animals. (default=0.47)}

\item{testSensitivityIh}{test sensitivity of highly positive animals. (default=0.71)}

\item{testSpecificity}{test specificity. (default=1.0)}

\item{startGrazing}{Grazing start week (from 0 to 51, default=14L, beginning of April)}

\item{endGrazing}{Grazing end week (from 0 to 51, default=46L, mid November)}

\item{ageAtWeaning}{From this age, in terms of number of weeks, the animal is weaned (default=10L)}

\item{ageAtGrazing}{From this age, in terms of number of weeks, the animal can go outside during grazing period (default=26L)}

\item{ageYoungHeifer}{From this age, in terms of number of weeks, the animal is considered to be a young heifer (default=52L, one year old)}

\item{ageHeifer}{From this age, in terms of number of weeks, the animal is considered to be a heifer, starts reproduction (default=91L)}

\item{ageAtFirstCalving}{From this age, in terms of number of weeks, the animal is considered to be a cow, calves for the first time (default=130L)}

\item{noInfection}{all herds will be initialized without infected animals, so no PTB spread (default=FALSE)}

\item{InitialInfection}{if not empty, then load initial 'infection' from this DF/tibble: 1 line = 1 infected herd with id, nbInfectedAnimals, prevalence and repetitionNum refering to a line in 'initCondFileName' (tibble/DF, default=0 rows)}

\item{BirthFemaleEvents}{a tibble/DF - see ?BirthFemaleEvents}

\item{CullingRates}{a tibble/DF - see ?CullingRates}

\item{Headcounts}{a tibble/DF - see ?Headcounts}

\item{Network}{a tibble/DF - see ?Network}

\item{seed}{the random number seed to use}
}
\value{
a list of tibbles corresponding to the files output by the model
}
\description{
This is a wrapper for the INRAE PTB model.  Note that (currently) this is a very thin wrapper; all RNG are C++ i.e. set.seed will have no effect.  Also there is no checking for interrupt signals...
}
